Create Special sum of array
Given an array of length N, which contains single digit elements at every index. You need to find and return the sum of all elements of the array. But the final sum should also be a single digit.
In order to keep the output single digit - you need to keep adding the output number digits till single digit is left.